Release builds for macos

+++ build_scripts/release_build.sh
@@ -15,37 +15,42 @@ error() {
     tput sgr0
 }
 
-# needed to keep user ID in and outside Docker in sync to be able to write to workspace directory
-uid="$(id -u)"
-image="$DIST":"$ARCH"-uid"$uid"
-dockerfile=containers/"$DIST"/Dockerfile."$ARCH"
+if [[ "$OSTYPE" == "darwin"* ]]; then
+    bash build_scripts/build-snowball.sh
+else
 
-echo $dockerfile
-echo $uid
-if [ ! -d "containers/$DIST" ]; then
-    error "Unknown dist: $DIST"
-    exit 2
+    # needed to keep user ID in and outside Docker in sync to be able to write to workspace directory
+    uid="$(id -u)"
+    image="$DIST":"$ARCH"-uid"$uid"
+    dockerfile=containers/"$DIST"/Dockerfile."$ARCH"
+
+    echo $dockerfile
+    echo $uid
+    if [ ! -d "containers/$DIST" ]; then
+        error "Unknown dist: $DIST"
+        exit 2
+    fi
+
+    if [ "$DIST" == "common" ]; then
+        error "\"common\" is not a distro"
+        exit 2
+    fi
+
+    if [ ! -f "$dockerfile" ]; then
+        error "Dockerfile $dockerfile could not be found"
+        exit 3
+    fi
+
+    export C="/usr/bin/g++"
+    export CXX="/usr/bin/g++-9"
+
+    # build image to cache dependencies
+    docker build -t "$image" -f "$dockerfile" --build-arg UID="$uid" containers/"$DIST"
+
+    # run build inside this image
+    EXTRA_ARGS=()
+    [ -t 1 ] && EXTRA_ARGS+=("-t")
+    docker run --rm -i "${EXTRA_ARGS[@]}" -e DOCKER=1 -v "$(readlink -f .)":/ws "$image" bash -xc "cd /ws && bash build_scripts/build-snowball.sh"
+
+    ls
 fi
-
-if [ "$DIST" == "common" ]; then
-    error "\"common\" is not a distro"
-    exit 2
-fi
-
-if [ ! -f "$dockerfile" ]; then
-    error "Dockerfile $dockerfile could not be found"
-    exit 3
-fi
-
-export C="/usr/bin/g++"
-export CXX="/usr/bin/g++-9"
-
-# build image to cache dependencies
-docker build -t "$image" -f "$dockerfile" --build-arg UID="$uid" containers/"$DIST"
-
-# run build inside this image
-EXTRA_ARGS=()
-[ -t 1 ] && EXTRA_ARGS+=("-t")
-docker run --rm -i "${EXTRA_ARGS[@]}" -e DOCKER=1 -v "$(readlink -f .)":/ws "$image" bash -xc "cd /ws && bash build_scripts/build-snowball.sh"
-
-ls
